src/pyproject.toml
[tool.poetry]
name = "lengua-sim"
version = "0.1.0"
description = ""
authors = ["Mike Zimin <mikeziminio@gmail.com>"]
readme = "../README.md"

[tool.poetry.dependencies]
python = "~3.11"
openai = "^1.7.0"
python-dotenv = "^1.0.0"
aiogram = "^3.3.0"

[tool.poetry.group.dev.dependencies]
chardet = "^5.2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
